+++ e2e/cmdenvvars/cmdenvvars.go
@@ -6,6 +6,7 @@
 package cmdenvvars
 
 import (
+	"fmt"
 	"os"
 	"path/filepath"
 	"testing"
@@ -259,6 +260,39 @@ func (c ctx) testSingularitySypgpDir(t *testing.T) {
 	}
 }
 
+func (c ctx) testSingularityConfigDir(t *testing.T) {
+	// Test plan:
+	//
+	// - create a temporary directory to be the configuration directory
+	// - run 'singularity remote list' to create the remote.yaml
+	//   file inside the configuration directory, with the
+	//   SINGULARITY_CONFIGDIR environment variables set to that temporary
+	//   directory
+	// - assert that the file has been created
+	//
+	// If the file is in the temporary directory, it means
+	// singularity followed the SINGULARITY_CONFIGDIR environment
+	// variable (checked in pkg/syfs) to set the configuration directory.
+
+	configDir, cleanup := setupTemporaryDir(t, c.env.TestDir, "config-dir")
+	defer cleanup(t)
+
+	environ := append(os.Environ(), fmt.Sprintf("SINGULARITY_CONFIGDIR=%s", configDir))
+	c.env.RunSingularity(
+		t,
+		e2e.WithProfile(e2e.UserProfile),
+		e2e.WithCommand("remote"),
+		e2e.WithArgs("list"),
+		e2e.WithEnv(environ),
+		e2e.ExpectExit(0),
+	)
+
+	remotePath := filepath.Join(configDir, "remote.yaml")
+	if _, err := os.Stat(remotePath); os.IsNotExist(err) {
+		t.Fatalf("failed to find remote.yaml (expected: %s)", remotePath)
+	}
+}
+
 // E2ETests is the main func to trigger the test suite
 func E2ETests(env e2e.TestEnv) testhelper.Tests {
 	c := ctx{
@@ -270,5 +304,6 @@ func E2ETests(env e2e.TestEnv) testhelper.Tests {
 		"SINGULARITY_CACHEDIR":      c.testSingularityCacheDir,
 		"singularity disable cache": c.testSingularityDisableCache,
 		"SINGULARITY_SYPGPDIR":      c.testSingularitySypgpDir,
+		"config directory":          c.testSingularityConfigDir,
 	}
 }