+++ pkg/syfs/syfs.go
@@ -23,6 +23,7 @@ const (
 	RemoteCache    = "remote-cache"
 	DockerConfFile = "docker-config.json"
 	singularityDir = ".singularity"
+	configDirEnv   = "SINGULARITY_CONFIGDIR"
 )
 
 // cache contains the information for the current user
@@ -43,6 +44,11 @@ func ConfigDir() string {
 }
 
 func configDir() string {
+	configDir := os.Getenv(configDirEnv)
+	if configDir != "" {
+		return configDir
+	}
+
 	user, err := user.Current()
 	if err != nil {
 		sylog.Warningf("Could not lookup the current user's information: %s", err)